refactor: remove subcomponents folder from Cart folder
